requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pyotp==2.9.0
aiohttp==3.8.5
smartapi-python==1.1.0
pycryptodome==3.21.0
matplotlib==3.8.1
logzero==1.7.0
websocket-client==1.8.0
openai==1.0.0
numpy==1.26.4